backend/database/PostgreSQL/init/requirements.txt
psycopg2-binary==2.9.9
yfinance==0.2.43
pandas==2.2.2
